random-xkcd: rename package to org.appdevforall.randomxkcd + author "App Dev For All"
Address Daniel-ADFA's PR #6 review comments:

- Move Kotlin source from `com.codeonthego.xkcdrandom` to
  `org.appdevforall.randomxkcd` to match the convention set by
  `ndk-installer-plugin` (the most recent precedent in the repo).
- Update `plugin.id` (`com.codeonthego.xkcdrandom` →
  `org.appdevforall.randomxkcd`), `plugin.main_class`,
  `applicationId` / `namespace` in `build.gradle.kts`, the
  ProGuard `-keep` rule, the Tier-3 walkthrough snippets in
  `assets/docs/index.html`, the README's source-layout block,
  and the host-served Tier-3 URL in the README.
- Change `plugin.author` from "Code on the Go Team" to
  "App Dev For All" (consistent with the other plugins, which
  use "App Dev for All" / "App Dev For All" — adopted Daniel's
  exact casing).

Verified: `./gradlew testDebugUnitTest` green; `./gradlew
clean assemblePluginDebug` produces a `.cgp` whose packaged
manifest reports the new id, author, and main_class via
`aapt2 dump xmltree`.
