+++ random-xkcd/proguard-rules.pro
@@ -1,4 +1,4 @@
 # Add project specific ProGuard rules here.
 
 # Keep plugin classes — the IDE loads them by reflection via plugin.main_class.
--keep class com.codeonthego.xkcdrandom.** { *; }
+-keep class org.appdevforall.randomxkcd.** { *; }
